extract websocket functionality to the connector
